feat(task): add resubmit functionality for rejected tasks and completion status tracking
- Introduced a new endpoint to allow users to resubmit rejected tasks, changing their status to PENDING.
- Added completion status tracking to task memberships, enabling better management of task states.
- Updated API documentation to reflect the new resubmit feature and completion status enumeration.
- Implemented necessary database migrations to support the new completion status field.
