+++ src/main/kotlin/org/rucca/cheese/task/TaskAIAdviceStatus.kt
@@ -1,0 +1,8 @@
+package org.rucca.cheese.task
+
+enum class TaskAIAdviceStatus {
+    PENDING,
+    PROCESSING,
+    COMPLETED,
+    FAILED,
+}
