+++ src/main/kotlin/org/rucca/cheese/task/TaskSubmissionEntry.kt
@@ -13,8 +13,6 @@ import jakarta.persistence.*
 import org.hibernate.annotations.SQLRestriction
 import org.rucca.cheese.attachment.Attachment
 import org.rucca.cheese.common.persistent.BaseEntity
-import org.rucca.cheese.common.persistent.IdType
-import org.springframework.data.jpa.repository.JpaRepository
 
 @Entity
 @SQLRestriction("deleted_at IS NULL")
@@ -27,7 +25,3 @@ class TaskSubmissionEntry(
     val contentText: String? = null,
     @ManyToOne(fetch = FetchType.LAZY) val contentAttachment: Attachment? = null, // nullable
 ) : BaseEntity()
-
-interface TaskSubmissionEntryRepository : JpaRepository<TaskSubmissionEntry, IdType> {
-    fun findAllByTaskSubmissionId(taskSubmissionId: IdType): List<TaskSubmissionEntry>
-}
