+++ src/main/kotlin/org/rucca/cheese/task/TaskSubmissionSchema.kt
@@ -1,0 +1,16 @@
+package org.rucca.cheese.task
+
+import jakarta.persistence.Column
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class TaskSubmissionSchema(
+    @Column(nullable = false) val index: Int? = null,
+    @Column(nullable = false) val description: String? = null,
+    @Column(nullable = false) val type: TaskSubmissionEntryType? = null,
+)
+
+enum class TaskSubmissionEntryType {
+    TEXT,
+    ATTACHMENT,
+}
