+++ src/main/kotlin/org/rucca/cheese/task/ai/TaskAIAdviceResponseProcessor.kt
@@ -1,4 +1,4 @@
-package org.rucca.cheese.task
+package org.rucca.cheese.task.ai
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
